feat: implement ProductPurchasePanel for enhanced cart functionality
- Added a new `ProductPurchasePanel` component to streamline the add-to-cart process.
- Moved cart-related logic from `ProductSummarySection` to the new panel for better separation of concerns.
- Updated the header cart button to open the Foxy Sidecart and document environment configurations.
- Adjusted the quantity handling and form submission to improve user experience and data integrity.
